Mini_project/requirements.txt
Flask==2.3.3
tensorflow==2.12.0
Pillow==10.0.1
numpy==1.23.5